login validacion de form

+++ src/pages/Admin.jsx
@@ -2,6 +2,8 @@ import { useEffect, useState } from 'react'
 import FormViaje from '../components/FormViaje'
 import '../css/admin.css'
 import FormEdit from '../components/FormEdit'
+import Swal from 'sweetalert2'
+import { useAuth } from '../context/AuthProvider'
 
 const Admin = () => {
 
@@ -12,7 +14,10 @@ const Admin = () => {
     const [viaje, setViaje] = useState({})
     const [edit, isEdit] = useState(false)
 
+    const { token, validateToken } = useAuth();
+
     useEffect(() => {
+        validateToken();
         getViajes()
     }, [])
 
@@ -20,7 +25,7 @@ const Admin = () => {
         fetch("http://localhost:8080/api/viajes", {
             method: "GET",
             headers: {
-                "content-type": "application/json"
+                "content-type": "application/json",
             }
         }).then(res => res.json())
             .then(data => setViajes(data))
@@ -30,33 +35,59 @@ const Admin = () => {
         setPopup(!popup)
     }
 
-    const handleEdit = id => {
-        isEdit(true)
-        fetch("http://localhost:8080/api/viajes/ " + id, {
-            method: "GET",
-            headers: {
-                "content-type": "application/json"
-            }
-        }).then(res => res.json())
-            .then(data => setViaje(data))
-            .then(() => togglePopup())
-    }
+    // const handleEdit = id => {
+    //     isEdit(true)
+    //     console.log(id);
+    //     fetch("http://localhost:8080/api/viajes/ " + id, {
+    //         method: "GET",
+    //         headers: {
+    //             "content-type": "application/json"
+    //         }
+    //     }).then(res => res.json())
+    //         .then(data => setViaje(data))
+    //         .then(() => togglePopup())
+    // }
 
     const handleDelete = id => {
-        fetch("http://localhost:8080/api/viajes/" + id, {
-            method: "DELETE"
-        }).then(() => getViajes())
+
+        Swal.fire({
+            title: "Seguro que desea eliminar el viaje?",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            confirmButtonText: "Si",
+            cancelButtonColor: "#d33",
+            cancelButtonText: "No",
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch("http://localhost:8080/api/viajes/" + id, {
+                    method: "DELETE",
+                    headers: {
+                        "authorization": "Bearer " + token,
+                        "content-type": "application/json"
+                    },
+                    credentials: "include"
+                }).then(() => getViajes())
+                Swal.fire({
+                    title: "Deleted!",
+                    text: "Your file has been deleted.",
+                    icon: "success"
+                });
+            }
+        });
+
     }
 
     const handleFavorite = (id) => {
-        fetch(`http://localhost:8080/api/viajes/favorito/${id}`, {
-            method: 'PUT',
+        fetch(`http://localhost:8080/api/viajes/favoritos/${id}`, {
+            method: 'POST',
             headers: {
+                "authorization": "Bearer " + token,
                 'Content-Type': 'application/json',
             },
         })
-        .then(() => getViajes()) // Actualizar la lista de viajes después de marcar/desmarcar como favorito
-        .catch((error) => console.error('Error al marcar/desmarcar como favorito:', error));
+            .then(() => getViajes()) // Actualizar la lista de viajes después de marcar/desmarcar como favorito
+            .catch((error) => console.error('Error al marcar/desmarcar como favorito:', error));
     };
 
     const closePopup = (e) => {
@@ -141,9 +172,9 @@ const Admin = () => {
                                                 <button onClick={() => handleDelete(d.id)}>
                                                     <img className='2xl:w-16 lg:w-8 md:w-6' src="./src/img/delete.svg" alt="" />
                                                 </button>
-                                                <button onClick={() => handleEdit(d.id)}>
+                                                {/* <button onClick={() => handleEdit(d.id)}>
                                                     <img className='2xl:w-16 lg:w-8 md:w-6' src="./src/img/edit.svg" alt="" />
-                                                </button>
+                                                </button> */}
                                             </div>
                                         </td>
                                     </tr>
@@ -156,7 +187,7 @@ const Admin = () => {
             {popup && (
                 <div onClick={closePopup} className='overlay'>
                     <div className=''>
-                        {edit ? <FormEdit popup={popup} onClose={onClose} viaje={viaje} /> : <FormViaje popup={popup} onClose={onClose} />}
+                        {edit ? <FormEdit popup={popup} onClose={onClose} viaje={viaje} setPopup={setPopup} /> : <FormViaje popup={popup} onClose={onClose} setPopup={setPopup} />}
                     </div>
                 </div>
             )
